requirements.txt
playwright>=1.49.0
pytest>=8.3.0
pytest-playwright>=0.6.2
pytest-html>=4.1.0
python-dotenv>=1.0.0
pytest-json-report>=1.5.0
